coldstoragefix/requirements.txt
# requirements.txt – Cold Storage IoT Monitoring System
# Install dengan: pip install -r requirements.txt


flask==3.1.0
flask-cors==5.0.0
pymongo==4.10.1
paho-mqtt==2.1.0
matplotlib==3.9.2